chore: moving UserResponseDto to dto package
