+++ api/src/main/java/lab/en2b/quizapi/game/dtos/GameResponseDto.java
@@ -2,14 +2,11 @@ package lab.en2b.quizapi.game.dtos;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
-import lab.en2b.quizapi.commons.user.UserResponseDto;
+import lab.en2b.quizapi.commons.user.dtos.UserResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
 
 @AllArgsConstructor
 @Data
