+++ api/src/test/java/lab/en2b/quizapi/game/GameServiceTest.java
@@ -2,7 +2,7 @@ package lab.en2b.quizapi.game;
 
 import ch.qos.logback.core.util.TimeUtil;
 import lab.en2b.quizapi.commons.user.User;
-import lab.en2b.quizapi.commons.user.UserResponseDto;
+import lab.en2b.quizapi.commons.user.dtos.UserResponseDto;
 import lab.en2b.quizapi.commons.user.UserService;
 import lab.en2b.quizapi.commons.user.mappers.UserResponseDtoMapper;
 import lab.en2b.quizapi.game.dtos.GameAnswerDto;
